crewai-stripe/pyproject.toml
[project]
name = "stripe-crew"
version = "0.1.0"
description = "Stripe integration using crewAI"
authors = [{ name = "Noah Cashin", email = "noahc959@icloud.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai[tools]>=0.86.0,<1.0.0",
    "stripe-agent-toolkit>=0.1.0",
    "python-dotenv>=1.0.0",
    "stripe>=7.0.0"
]

[project.scripts]
stripe-crew = "src.stripe.main:run"
run_crew = "src.stripe.main:run"
train = "src.stripe.main:train"
replay = "src.stripe.main:replay"
test = "src.stripe.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
include = [
    "src/**/*.py",
    "src/**/*.yaml"
]
